User controller updated (authorization)

+++ src/controller/userController.js
@@ -8,13 +8,25 @@ import jwt from "jsonwebtoken";
 const JWT_SECRET = process.env.JWT_SECRET;
 class Controller {
   hashPassword(password) {
-    const salt = bcrypt.genSaltSync(10);
-    const hash = bcrypt.hashSync(password, salt);
+    // I can pass an undefined password in case i don't update it, so use try and catch
+    try {
+      const salt = bcrypt.genSaltSync(10);
+      const hash = bcrypt.hashSync(password, salt);
 
-    return hash;
+      return hash;
+    } catch (err) {
+      console.error(err);
+      // password is undefined if it got here
+      return undefined;
+    }
   }
 
   async index(req, res) {
+    if (req.loggedUser.role !== "Administrator")
+      return res
+        .status(403)
+        .json({ message: "You are not allowed to see all users." }); // 403 Forbidden
+    // only see all users if it is the administrator
     const users = await UserModel.find().lean();
     res.json({ users });
   }
@@ -42,8 +54,23 @@ class Controller {
     try {
       const user = await UserModel.findById(id);
 
+      // if user exists allow this request only if the current user logged is an administrator or
+      // the user has the same id as the specified in params
       if (user) {
-        return res.json({ user });
+        // user._id.equals(string) = user._id is an instance of ObjectId, 
+        // to campare their values it's necessary to use the ObjectID.equals method
+        // or document._id.toString() can also works
+        // or id === req.loggedUser._id = because both are string 
+        if ( req.loggedUser.role === "Administrator" ||
+          user._id.equals(req.loggedUser._id) ) { 
+          return res.json({ user });
+        }
+
+        return res
+          .status(403)
+          .json({
+            message: "You are not allowed to see this user's information.",
+          });
       }
 
       res.status(404).json({ message: "User not found!" });
@@ -77,8 +104,14 @@ class Controller {
 
   async update(req, res) {
     const { id } = req.params;
+
+    // only allow update if logged user has the same id as the specified in params
+    if (id !== req.loggedUser._id)
+      return res.status(403).json({message: "You are not allowed to update this user's information"});
+
     try {
       const { email, password, phones, name, role } = req.body;
+
       const user = await UserModel.findByIdAndUpdate(
         id,
         { email, name, password: this.hashPassword(password), phones, role },
@@ -97,6 +130,11 @@ class Controller {
 
   async remove(req, res) {
     const id = req.params.id;
+
+    // only allow update if logged user has the same id as the specified in params
+    if (id !== req.loggedUser._id)
+      return res.status(403).json({message: "You are not allowed to delete this user"});
+
     try {
       const user = await UserModel.findByIdAndDelete(id);
